requirements.txt
streamlit==1.34.0
gunicorn==22.0.0
Jinja2==3.1.3
python-dotenv==1.0.1
langchain-chroma==0.1.2
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
openai==1.21.2
pdf2image==1.17.0
docx2txt==0.8
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.15.1
pinecone-client==3.2.2
pydantic==2.7.0
pydantic_core==2.18.1
pypdf==4.2.0
pypdfium2==4.29.0
python-docx==1.1.0
requests==2.31.0
tiktoken==0.6.0
tokenizers==0.15.2
transformers==4.39.3
unstructured==0.13.2
unstructured-client==0.18.0
unstructured-inference==0.7.25
unstructured.pytesseract==0.3.12
